Add files via upload
The program will take input in the form of the amount in one currency and the desired currency to convert to. It will then retrieve the exchange rate between the two currencies and perform the conversion. The converted amount will be displayed as output. By building a currency converter using Python, you can easily calculate and compare currency values for various purposes such as travel, online shopping, or financial analysis.