Add files via upload
In this Python implementation, you'll learn how to create a simple version of the game using the Pygame library. The player will use arrow keys to navigate the snake, and the game will keep track of the score as the snake grows. It's a fun and interactive project that showcases the fundamentals of game development in Python.